+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Switches to gobble for building
+* Makes labels and values bindable
+* Refactor to work with Ractive edge as of 2014-12-26
+
 ## 0.1.0
 
 * __BUG:__ Prevents double toggle when click event doesn't originate in the button.